Time: 0 ms (100.00%) | Memory: 42 MB (80.84%) - LeetSync

+++ 50-powx-n/powx-n.java
@@ -22,7 +22,7 @@ class Solution {
         double res = 0;
 
         if(n >= 0){
-            long m = n;
+            long m = (long)n;
             res = getPower(x, m);
         }
         else{
